// webapp/src/app/enums/built-in-ai-api.enum.ts
export enum BuiltInAiApi {
  LanguageDetector = "Language Detector",
  Translator = "Translator",
  Summarizer = "Summarizer",
  Prompt = "Prompt",
  PromptWithImage = "Prompt (with Image)",
  PromptWithAudio = "Prompt (with Audio)",
  Proofreader = "Proofreader",
  Rewriter = "Rewriter",
  Writer = "Writer",
}